@@ -7,133 +7,88 @@ import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
 
-import 'package:args/args.dart';
 import 'package:dir_contents_diff/dir_contents_diff.dart' show dirContentsDiff;
 import 'package:engine_repo_tools/engine_repo_tools.dart';
 import 'package:path/path.dart';
 import 'package:process/process.dart';
 import 'package:skia_gold_client/skia_gold_client.dart';
 
+import 'utils/adb_logcat_filtering.dart';
+import 'utils/environment.dart';
 import 'utils/logs.dart';
+import 'utils/options.dart';
 import 'utils/process_manager_extension.dart';
 import 'utils/screenshot_transformer.dart';
 
-void _withTemporaryCwd(String path, void Function() callback) {
-  final String originalCwd = Directory.current.path;
-  Directory.current = Directory(path).parent.path;
-
-  try {
-    callback();
-  } finally {
-    Directory.current = originalCwd;
-  }
-}
-
 // If you update the arguments, update the documentation in the README.md file.
 void main(List<String> args) async {
-  final Engine? engine = Engine.tryFindWithin();
-  final ArgParser parser = ArgParser()
-    ..addFlag(
-      'help',
-      help: 'Prints usage information',
-      negatable: false,
-    )
-    ..addOption(
-      'adb',
-      help: 'Absolute path to the adb tool',
-      defaultsTo: engine != null ? join(
-        engine.srcDir.path,
-        'third_party',
-        'android_tools',
-        'sdk',
-        'platform-tools',
-        'adb',
-      ) : null,
-    )
-    ..addOption(
-      'out-dir',
-      help: 'Out directory',
-      defaultsTo:
-        engine?.
-        outputs().
-        where((Output o) => basename(o.path.path).startsWith('android_')).
-        firstOrNull?.
-        path.path,
-    )
-    ..addOption(
-      'smoke-test',
-      help: 'runs a single test to verify the setup',
-    )
-    ..addFlag(
-      'use-skia-gold',
-      help: 'Use Skia Gold to compare screenshots.',
-      defaultsTo: isLuciEnv,
-    )
-    ..addFlag(
-      'enable-impeller',
-      help: 'Enable Impeller for the Android app.',
-    )
-    ..addOption('output-contents-golden',
-      help:
-        'Path to a file that will be used to check the contents of the output to make sure everything was created.')
-    ..addOption(
-      'impeller-backend',
-      help: 'The Impeller backend to use for the Android app.',
-      allowed: <String>['vulkan', 'opengles'],
-      defaultsTo: 'vulkan',
-    )
-    ..addOption(
-      'logs-dir',
-      help: 'The directory to store the logs and screenshots. Defaults to '
-            'the value of the FLUTTER_LOGS_DIR environment variable, if set, '
-            'otherwise it defaults to a path within out-dir.',
-      defaultsTo: Platform.environment['FLUTTER_LOGS_DIR'],
-    );
+  // Get some basic environment information to guide the rest of the program.
+  final Environment environment = Environment(
+    isCi: Platform.environment['LUCI_CONTEXT'] != null,
+    showVerbose: Options.showVerbose(args),
+    logsDir: Platform.environment['FLUTTER_LOGS_DIR'],
+  );
 
+  // Determine if the CWD is within an engine checkout.
+  final Engine? localEngineDir = Engine.tryFindWithin();
+
+  // Show usage if requested.
+  if (Options.showUsage(args)) {
+    stdout.writeln(Options.usage(
+      environment: environment,
+      localEngineDir: localEngineDir,
+    ));
+    return;
+  }
+
+  // Parse the command line arguments.
+  final Options options;
+  try {
+    options = Options.parse(
+      args,
+      environment: environment,
+      localEngine: localEngineDir,
+    );
+  } on FormatException catch (error) {
+    stderr.writeln(error);
+    stderr.writeln(Options.usage(
+      environment: environment,
+      localEngineDir: localEngineDir,
+    ));
+    exitCode = 1;
+    return;
+  }
+
+  // Capture CTRL-C.
+  late final StreamSubscription<void> onSigint;
   runZonedGuarded(
     () async {
-      final ArgResults results = parser.parse(args);
-      if (results['help'] as bool) {
-        stdout.writeln(parser.usage);
-        return;
-      }
-
-      if (results['out-dir'] == null) {
-        panic(<String>['--out-dir is required']);
-      }
-      if (results['adb'] == null) {
-        panic(<String>['--adb is required']);
-      }
-
-      final Directory outDir = Directory(results['out-dir'] as String);
-      final File adb = File(results['adb'] as String);
-      final bool useSkiaGold = results['use-skia-gold'] as bool;
-      final String? smokeTest = results['smoke-test'] as String?;
-      final bool enableImpeller = results['enable-impeller'] as bool;
-      final String? contentsGolden = results['output-contents-golden'] as String?;
-      final _ImpellerBackend? impellerBackend = _ImpellerBackend.tryParse(results['impeller-backend'] as String?);
-      if (enableImpeller && impellerBackend == null) {
-        panic(<String>['invalid graphics-backend', results['impeller-backend'] as String? ?? '<null>']);
-      }
-      final Directory logsDir = Directory(results['logs-dir'] as String? ?? join(outDir.path, 'scenario_app', 'logs'));
+      onSigint = ProcessSignal.sigint.watch().listen((_) {
+        onSigint.cancel();
+        panic(<String>['Received SIGINT']);
+      });
       await _run(
-        outDir: outDir,
-        adb: adb,
-        smokeTestFullPath: smokeTest,
-        useSkiaGold: useSkiaGold,
-        enableImpeller: enableImpeller,
-        impellerBackend: impellerBackend,
-        logsDir: logsDir,
-        contentsGolden: contentsGolden,
+        verbose: options.verbose,
+        outDir: Directory(options.outDir),
+        adb: File(options.adb),
+        smokeTestFullPath: options.smokeTest,
+        useSkiaGold: options.useSkiaGold,
+        enableImpeller: options.enableImpeller,
+        impellerBackend: _ImpellerBackend.tryParse(options.impellerBackend),
+        logsDir: Directory(options.logsDir),
+        contentsGolden: options.outputContentsGolden,
+        ndkStack: options.ndkStack,
       );
+      onSigint.cancel();
       exit(0);
     },
     (Object error, StackTrace stackTrace) {
+      onSigint.cancel();
       if (error is! Panic) {
-        stderr.writeln(error);
+        stderr.writeln('Unhandled error: $error');
         stderr.writeln(stackTrace);
       }
-      exit(1);
+      exitCode = 1;
     },
   );
 }
@@ -155,6 +110,7 @@ enum _ImpellerBackend {
 }
 
 Future<void> _run({
+  required bool verbose,
   required Directory outDir,
   required File adb,
   required String? smokeTestFullPath,
@@ -163,21 +119,16 @@ Future<void> _run({
   required _ImpellerBackend? impellerBackend,
   required Directory logsDir,
   required String? contentsGolden,
+  required String ndkStack,
 }) async {
   const ProcessManager pm = LocalProcessManager();
-
-  if (!outDir.existsSync()) {
-    panic(<String>['out-dir does not exist: $outDir', 'make sure to build the selected engine variant']);
-  }
-
-  if (!adb.existsSync()) {
-    panic(<String>['cannot find adb: $adb', 'make sure to run gclient sync']);
-  }
-
   final String scenarioAppPath = join(outDir.path, 'scenario_app');
   final String logcatPath = join(logsDir.path, 'logcat.txt');
 
   // TODO(matanlurey): Use screenshots/ sub-directory (https://github.com/flutter/flutter/issues/143604).
+  if (!logsDir.existsSync()) {
+    logsDir.createSync(recursive: true);
+  }
   final String screenshotPath = logsDir.path;
   final String apkOutPath = join(scenarioAppPath, 'app', 'outputs', 'apk');
   final File testApk = File(join(apkOutPath, 'androidTest', 'debug', 'app-debug-androidTest.apk'));
@@ -185,11 +136,17 @@ Future<void> _run({
   log('writing logs and screenshots to ${logsDir.path}');
 
   if (!testApk.existsSync()) {
-    panic(<String>['test apk does not exist: ${testApk.path}', 'make sure to build the selected engine variant']);
+    panic(<String>[
+      'test apk does not exist: ${testApk.path}',
+      'make sure to build the selected engine variant'
+    ]);
   }
 
   if (!appApk.existsSync()) {
-    panic(<String>['app apk does not exist: ${appApk.path}', 'make sure to build the selected engine variant']);
+    panic(<String>[
+      'app apk does not exist: ${appApk.path}',
+      'make sure to build the selected engine variant'
+    ]);
   }
 
   // Start a TCP socket in the host, and forward it to the device that runs the tests.
@@ -197,17 +154,25 @@ Future<void> _run({
   // for the screenshots.
   // On LUCI, the host uploads the screenshots to Skia Gold.
   SkiaGoldClient? skiaGoldClient;
-  late  ServerSocket server;
+  late final ServerSocket server;
   final List<Future<void>> pendingComparisons = <Future<void>>[];
+  final List<Socket> pendingConnections = <Socket>[];
   await step('Starting server...', () async {
     server = await ServerSocket.bind(InternetAddress.anyIPv4, _tcpPort);
-    stdout.writeln('listening on host ${server.address.address}:${server.port}');
+    if (verbose) {
+      stdout.writeln('listening on host ${server.address.address}:${server.port}');
+    }
     server.listen((Socket client) {
-      stdout.writeln('client connected ${client.remoteAddress.address}:${client.remotePort}');
+      if (verbose) {
+        stdout.writeln('client connected ${client.remoteAddress.address}:${client.remotePort}');
+      }
+      pendingConnections.add(client);
       client.transform(const ScreenshotBlobTransformer()).listen((Screenshot screenshot) {
         final String fileName = screenshot.filename;
         final Uint8List fileContent = screenshot.fileContent;
-        log('host received ${fileContent.lengthInBytes} bytes for screenshot `$fileName`');
+        if (verbose) {
+          log('host received ${fileContent.lengthInBytes} bytes for screenshot `$fileName`');
+        }
         assert(skiaGoldClient != null, 'expected Skia Gold client');
         late File goldenFile;
         try {
@@ -215,26 +180,26 @@ Future<void> _run({
         } on FileSystemException catch (err) {
           panic(<String>['failed to create screenshot $fileName: $err']);
         }
-        log('wrote ${goldenFile.absolute.path}');
+        if (verbose) {
+          log('wrote ${goldenFile.absolute.path}');
+        }
         if (isSkiaGoldClientAvailable) {
           final Future<void> comparison = skiaGoldClient!
-            .addImg(fileName, goldenFile,
-                    screenshotSize: screenshot.pixelCount)
-            .catchError((dynamic err) {
-              panic(<String>['skia gold comparison failed: $err']);
-            });
+              .addImg(fileName, goldenFile, screenshotSize: screenshot.pixelCount)
+              .catchError((dynamic err) {
+            panic(<String>['skia gold comparison failed: $err']);
+          });
           pendingComparisons.add(comparison);
         }
-      },
-      onError: (dynamic err) {
-        panic(<String>['error while receiving bytes: $err']);
-      },
-      cancelOnError: true);
+      }, onDone: () {
+        pendingConnections.remove(client);
+      });
     });
   });
 
   late Process logcatProcess;
   late Future<int> logcatProcessExitCode;
+  bool seenImpeller = false;
 
   final IOSink logcat = File(logcatPath).openWrite();
   try {
@@ -247,8 +212,48 @@ Future<void> _run({
       if (exitCode != 0) {
         panic(<String>['could not clear logs']);
       }
+
       logcatProcess = await pm.start(<String>[adb.path, 'logcat', '-T', '1']);
-      logcatProcessExitCode = pipeProcessStreams(logcatProcess, out: logcat);
+      final (Future<int> logcatExitCode, Stream<String> logcatOutput) = getProcessStreams(logcatProcess);
+
+      logcatProcessExitCode = logcatExitCode;
+      String? filterProcessId;
+
+      logcatOutput.listen((String line) {
+        // Always write to the full log.
+        logcat.writeln(line);
+        if (enableImpeller && !seenImpeller) {
+          seenImpeller = line.contains('Using the Impeller rendering backend');
+        }
+
+        // Conditionally parse and write to stderr.
+        final AdbLogLine? adbLogLine = AdbLogLine.tryParse(line);
+        if (verbose || adbLogLine == null) {
+          log(line);
+          return;
+        }
+
+        // If we haven't already found a process ID, try to find one.
+        // The process ID will help us filter out logs from other processes.
+        filterProcessId ??= adbLogLine.tryParseProcess();
+
+        // If this is a "verbose" log, possibly skip it.
+        final bool isVerbose = adbLogLine.isVerbose(filterProcessId: filterProcessId);
+        if (isVerbose || filterProcessId == null) {
+          // We've requested verbose output, so print everything.
+          if (verbose) {
+            adbLogLine.printFormatted();
+          }
+          return;
+        }
+
+        // It's a non-verbose log, so print it.
+        adbLogLine.printFormatted();
+      }, onError: (Object? err) {
+        if (verbose) {
+          logWarning('logcat stream error: $err');
+        }
+      });
     });
 
     await step('Configuring emulator...', () async {
@@ -279,10 +284,7 @@ Future<void> _run({
       log('using dimensions: ${json.encode(dimensions)}');
       skiaGoldClient = SkiaGoldClient(
         outDir,
-        dimensions: <String, String>{
-          'AndroidAPILevel': connectedDeviceAPILevel,
-          'GraphicsBackend': enableImpeller ? 'impeller-${impellerBackend!.name}' : 'skia',
-        },
+        dimensions: dimensions,
       );
     });
 
@@ -329,11 +331,11 @@ Future<void> _run({
         '-w',
         if (smokeTestFullPath != null)
           '-e class $smokeTestFullPath',
-        'dev.flutter.scenarios.test/dev.flutter.TestRunner',
         if (enableImpeller)
-          '-e enable-impeller',
+          '-e enable-impeller true',
         if (impellerBackend != null)
           '-e impeller-backend ${impellerBackend.name}',
+        'dev.flutter.scenarios.test/dev.flutter.TestRunner',
       ]);
       if (exitCode != 0) {
         panic(<String>['instrumented tests failed to run']);
@@ -348,29 +350,14 @@ Future<void> _run({
     });
   } finally {
     await server.close();
+    for (final Socket client in pendingConnections.toList()) {
+      client.close();
+    }
 
-    await step('Remove reverse port...', () async {
-      final int exitCode = await pm.runAndForward(<String>[
-        adb.path,
-        'reverse',
-        '--remove', 'tcp:3000',
-      ]);
+    await step('Killing test app and test runner...', () async {
+      final int exitCode = await pm.runAndForward(<String>[adb.path, 'shell', 'am', 'force-stop', 'dev.flutter.scenarios']);
       if (exitCode != 0) {
-        panic(<String>['could not unforward port']);
-      }
-    });
-
-    await step('Uninstalling app APK...', () async {
-      final int exitCode = await pm.runAndForward(<String>[adb.path, 'uninstall', 'dev.flutter.scenarios']);
-      if (exitCode != 0) {
-        panic(<String>['could not uninstall app apk']);
-      }
-    });
-
-    await step('Uninstalling test APK...', () async {
-      final int exitCode = await pm.runAndForward(<String>[adb.path, 'uninstall', 'dev.flutter.scenarios.test']);
-      if (exitCode != 0) {
-        panic(<String>['could not uninstall app apk']);
+        panic(<String>['could not kill test app']);
       }
     });
 
@@ -380,6 +367,65 @@ Future<void> _run({
       await logcatProcessExitCode;
     });
 
+    await step('Flush logcat...', () async {
+      await logcat.flush();
+      await logcat.close();
+      log('wrote logcat to $logcatPath');
+    });
+
+    if (enableImpeller) {
+      await step('Validating Impeller...', () async {
+        if (!seenImpeller) {
+          panic(<String>[
+            '--enable-impeller was specified, but Impeller was not used.',
+          ]);
+        }
+      });
+    }
+
+    await step('Symbolize stack traces', () async {
+      final ProcessResult result = await pm.run(
+        <String>[
+          ndkStack,
+          '-sym',
+          outDir.path,
+          '-dump',
+          logcatPath,
+        ],
+      );
+      if (result.exitCode != 0) {
+        panic(<String>['Failed to symbolize stack traces']);
+      }
+    });
+
+    await step('Remove reverse port...', () async {
+      final int exitCode = await pm.runAndForward(<String>[
+        adb.path,
+        'reverse',
+        '--remove',
+        'tcp:3000',
+      ]);
+      if (exitCode != 0) {
+        panic(<String>['could not unforward port']);
+      }
+    });
+
+    await step('Uninstalling app APK...', () async {
+      final int exitCode = await pm.runAndForward(
+          <String>[adb.path, 'uninstall', 'dev.flutter.scenarios']);
+      if (exitCode != 0) {
+        panic(<String>['could not uninstall app apk']);
+      }
+    });
+
+    await step('Uninstalling test APK...', () async {
+      final int exitCode = await pm.runAndForward(
+          <String>[adb.path, 'uninstall', 'dev.flutter.scenarios.test']);
+      if (exitCode != 0) {
+        panic(<String>['could not uninstall app apk']);
+      }
+    });
+
     await step('Wait for Skia gold comparisons...', () async {
       await Future.wait(pendingComparisons);
     });
@@ -387,19 +433,28 @@ Future<void> _run({
     if (contentsGolden != null) {
       // Check the output here.
       await step('Check output files...', () async {
+        // TODO(matanlurey): Resolve this in a better way. On CI this file always exists.
+        File(join(screenshotPath, 'noop.txt')).writeAsStringSync('');
         // TODO(gaaclarke): We should move this into dir_contents_diff.
-        _withTemporaryCwd(contentsGolden, () {
-          final int exitCode = dirContentsDiff(basename(contentsGolden), screenshotPath);
+        final String diffScreenhotPath = absolute(screenshotPath);
+        _withTemporaryCwd(absolute(dirname(contentsGolden)), () {
+          final int exitCode = dirContentsDiff(basename(contentsGolden), diffScreenhotPath);
           if (exitCode != 0) {
             panic(<String>['Output contents incorrect.']);
           }
         });
       });
     }
+  }
+}
 
-    await step('Flush logcat...', () async {
-      await logcat.flush();
-      await logcat.close();
-    });
+void _withTemporaryCwd(String path, void Function() callback) {
+  final String originalCwd = Directory.current.path;
+  Directory.current = Directory(path).path;
+
+  try {
+    callback();
+  } finally {
+    Directory.current = originalCwd;
   }
 }
