@@ -24,6 +24,7 @@
 
 #include "third_party/skia/include/core/SkBBHFactory.h"
 #include "third_party/skia/include/core/SkColorFilter.h"
+#include "third_party/skia/include/core/SkColorSpace.h"
 #include "third_party/skia/include/core/SkFontMgr.h"
 #include "third_party/skia/include/core/SkPictureRecorder.h"
 #include "third_party/skia/include/core/SkStream.h"
@@ -4528,10 +4529,14 @@ class DisplayListNopTest : public DisplayListRendering {
 
     auto sk_mode = static_cast<SkBlendMode>(mode);
     auto sk_color_filter = SkColorFilters::Blend(ToSk(color), sk_mode);
+    auto srgb = SkColorSpace::MakeSRGB();
     int all_flags = 0;
     if (sk_color_filter) {
       for (DlColor dst_color : test_dst_colors) {
-        DlColor result = DlColor(sk_color_filter->filterColor(ToSk(dst_color)));
+        SkColor4f dst_color_f = SkColor4f::FromColor(ToSk(dst_color));
+        DlColor result = DlColor(
+            sk_color_filter->filterColor4f(dst_color_f, srgb.get(), srgb.get())
+                .toSkColor());
         all_flags |= check_color_result(dst_color, result, dl, desc);
       }
       if ((all_flags & kWasMTB) != 0) {
