@@ -154,9 +154,8 @@ bool TiledTextureContents::Render(const ContentContext& renderer,
       [&renderer, &pipeline_method](ContentContextOptions options) {
         return (renderer.*pipeline_method)(options);
       };
-  return ColorSourceContents::DrawPositionsAndUVs<VS>(
-      Rect::MakeSize(texture_size), GetInverseEffectTransform(), renderer,
-      entity, pass, pipeline_callback, frame_info,
+  return ColorSourceContents::DrawGeometry<VS>(
+      renderer, entity, pass, pipeline_callback, frame_info,
       [this, &renderer, &is_external_texture,
        &uses_emulated_tile_mode](RenderPass& pass) {
         auto& host_buffer = renderer.GetTransientsBuffer();
@@ -215,7 +214,10 @@ bool TiledTextureContents::Render(const ContentContext& renderer,
         }
 
         return true;
-      });
+      },
+      /*enable_uvs=*/true,
+      /*texture_coverage=*/Rect::MakeSize(texture_size),
+      /*effect_transform=*/GetInverseEffectTransform());
 }
 
 std::optional<Snapshot> TiledTextureContents::RenderToSnapshot(
