@@ -10,6 +10,13 @@ mergeInto(LibraryManager.library, {
   $skwasm_support_setup: function() {
     const handleToCanvasMap = new Map();
     const associatedObjectsMap = new Map();
+
+    // This value represents the difference between the time origin of the main
+    // thread and whichever web worker this code is running on. This is so that
+    // when we report frame timings, that they are in the same time domain
+    // regardless of whether they are captured on the main thread or the web
+    // worker.
+    let timeOriginDelta;
     _skwasm_setAssociatedObjectOnThread = function(threadId, pointer, object) {
       PThread.pthreads[threadId].postMessage({
         skwasmMessage: 'setAssociatedObject',
@@ -20,6 +27,12 @@ mergeInto(LibraryManager.library, {
     _skwasm_getAssociatedObject = function(pointer) {
       return associatedObjectsMap.get(pointer);
     };
+    _skwasm_syncTimeOriginForThread = function(threadId) {
+      PThread.pthreads[threadId].postMessage({
+        skwasmMessage: 'syncTimeOrigin',
+        timeOrigin: performance.timeOrigin,
+      });
+    }
     _skwasm_registerMessageListener = function(threadId) {
       const eventListener = function({data}) {
         const skwasmMessage = data.skwasmMessage;
@@ -27,8 +40,16 @@ mergeInto(LibraryManager.library, {
           return;
         }
         switch (skwasmMessage) {
+          case 'syncTimeOrigin':
+            timeOriginDelta = performance.timeOrigin - data.timeOrigin;
+            return;
           case 'renderPictures':
-            _surface_renderPicturesOnWorker(data.surface, data.pictures, data.pictureCount, data.callbackId, performance.now());
+            _surface_renderPicturesOnWorker(
+              data.surface,
+              data.pictures,
+              data.pictureCount,
+              data.callbackId,
+              performance.now() + timeOriginDelta);
             return;
           case 'onRenderComplete':
             _surface_onRenderComplete(
@@ -101,7 +122,7 @@ mergeInto(LibraryManager.library, {
     };
     _skwasm_resolveAndPostImages = async function(surfaceHandle, imagePromises, rasterStart, callbackId) {
       const imageBitmaps = imagePromises ? await Promise.all(imagePromises) : [];
-      const rasterEnd = performance.now();
+      const rasterEnd = performance.now() + timeOriginDelta;
       postMessage({
         skwasmMessage: 'onRenderComplete',
         surface: surfaceHandle,
@@ -139,6 +160,8 @@ mergeInto(LibraryManager.library, {
   skwasm_getAssociatedObject__deps: ['$skwasm_support_setup'],
   skwasm_disposeAssociatedObjectOnThread: function () {},
   skwasm_disposeAssociatedObjectOnThread__deps: ['$skwasm_support_setup'],
+  skwasm_syncTimeOriginForThread: function() {},
+  skwasm_syncTimeOriginForThread__deps: ['$skwasm_support_setup'],
   skwasm_registerMessageListener: function() {},
   skwasm_registerMessageListener__deps: ['$skwasm_support_setup'],
   skwasm_dispatchRenderPictures: function() {},
