@@ -12,7 +12,7 @@ class MockAPKAssetProviderImpl : public APKAssetProviderInternal {
               (const, override));
 };
 
-TEST(APKAssetProvider, Clone) {
+TEST(APKAssetProvider, CloneAndEquals) {
   auto first_provider = std::make_unique<APKAssetProvider>(
       std::make_shared<MockAPKAssetProviderImpl>());
   auto second_provider = std::make_unique<APKAssetProvider>(
@@ -21,6 +21,8 @@ TEST(APKAssetProvider, Clone) {
 
   ASSERT_NE(first_provider->GetImpl(), second_provider->GetImpl());
   ASSERT_EQ(first_provider->GetImpl(), third_provider->GetImpl());
+  ASSERT_FALSE(*first_provider == *second_provider);
+  ASSERT_TRUE(*first_provider == *third_provider);
 }
 }  // namespace testing
 }  // namespace flutter
