@@ -9,7 +9,6 @@ import 'dart:io' as io;
 import 'package:engine_build_configs/engine_build_configs.dart';
 import 'package:engine_repo_tools/engine_repo_tools.dart';
 import 'package:engine_tool/src/commands/command_runner.dart';
-import 'package:engine_tool/src/commands/flags.dart';
 import 'package:engine_tool/src/environment.dart';
 import 'package:engine_tool/src/logger.dart';
 import 'package:litetest/litetest.dart';
@@ -32,17 +31,20 @@ void main() {
 
   final BuildConfig linuxTestConfig = BuildConfig.fromJson(
     path: 'ci/builders/linux_test_config.json',
-    map: convert.jsonDecode(fixtures.testConfig('Linux')) as Map<String, Object?>,
+    map: convert.jsonDecode(fixtures.testConfig('Linux'))
+        as Map<String, Object?>,
   );
 
   final BuildConfig macTestConfig = BuildConfig.fromJson(
     path: 'ci/builders/mac_test_config.json',
-    map: convert.jsonDecode(fixtures.testConfig('Mac-12')) as Map<String, Object?>,
+    map: convert.jsonDecode(fixtures.testConfig('Mac-12'))
+        as Map<String, Object?>,
   );
 
   final BuildConfig winTestConfig = BuildConfig.fromJson(
     path: 'ci/builders/win_test_config.json',
-    map: convert.jsonDecode(fixtures.testConfig('Windows-11')) as Map<String, Object?>,
+    map: convert.jsonDecode(fixtures.testConfig('Windows-11'))
+        as Map<String, Object?>,
   );
 
   final Map<String, BuildConfig> configs = <String, BuildConfig>{
@@ -76,7 +78,8 @@ void main() {
       configs: configs,
     );
     final int result = await runner.run(<String>[
-      'query', 'builds',
+      'query',
+      'builders',
     ]);
     expect(result, equals(0));
     expect(
@@ -92,7 +95,8 @@ void main() {
     );
   });
 
-  test('query command with --builder returns only from the named builder.', () async {
+  test('query command with --builder returns only from the named builder.',
+      () async {
     final Logger logger = Logger.test();
     final Environment env = linuxEnv(logger);
     final ToolCommandRunner runner = ToolCommandRunner(
@@ -100,18 +104,20 @@ void main() {
       configs: configs,
     );
     final int result = await runner.run(<String>[
-      'query', 'builds', '--$builderFlag', 'linux_test_config',
+      'query',
+      'builders',
+      '--builder',
+      'linux_test_config',
     ]);
     expect(result, equals(0));
     expect(
-      stringsFromLogs(logger.testLogs),
-      equals(<String>[
-        'Add --verbose to see detailed information about each builder\n',
-        '\n',
-        '"linux_test_config" builder:\n',
-        '   "build_name" config\n',
-      ]),
-    );
+        stringsFromLogs(logger.testLogs),
+        equals(<String>[
+          'Add --verbose to see detailed information about each builder\n',
+          '\n',
+          '"linux_test_config" builder:\n',
+          '   "build_name" config\n',
+        ]));
   });
 
   test('query command with --all returns all builds.', () async {
@@ -122,7 +128,9 @@ void main() {
       configs: configs,
     );
     final int result = await runner.run(<String>[
-      'query', 'builds', '--$allFlag',
+      'query',
+      'builders',
+      '--all',
     ]);
     expect(result, equals(0));
     expect(
