@@ -53,6 +53,15 @@ final Set<String> skippedPaths = <String>{
   r'flutter/third_party/gn',
   r'flutter/third_party/harfbuzz/docs',
   r'flutter/third_party/harfbuzz/util', // utils are command line tools that do not end up in the binary
+  r'flutter/third_party/icu/filters',
+  r'flutter/third_party/icu/fuzzers',
+  r'flutter/third_party/icu/scripts',
+  r'flutter/third_party/icu/source/common/unicode/uvernum.h', // this file contains strings that confuse the analysis
+  r'flutter/third_party/icu/source/config',
+  r'flutter/third_party/icu/source/data/brkitr/dictionaries/burmesedict.txt', // explicitly handled by ICU license
+  r'flutter/third_party/icu/source/data/brkitr/dictionaries/cjdict.txt', // explicitly handled by ICU license
+  r'flutter/third_party/icu/source/data/brkitr/dictionaries/laodict.txt', // explicitly handled by ICU license
+  r'flutter/third_party/icu/source/data/dtd',
   r'flutter/third_party/imgui',
   r'flutter/third_party/inja/doc', // documentation
   r'flutter/third_party/inja/third_party/amalgamate', // only used at build time
@@ -81,6 +90,7 @@ final Set<String> skippedPaths = <String>{
   r'flutter/third_party/pkg/process',
   r'flutter/third_party/pkg/process_runner',
   r'flutter/third_party/pkg/vector_math',
+  r'flutter/third_party/protobuf', // build-time dependency only
   r'flutter/third_party/pyyaml', // build-time dependency only
   r'flutter/third_party/rapidjson/contrib', // contains nothing that ends up in the binary executable
   r'flutter/third_party/rapidjson/doc', // documentation
@@ -161,15 +171,6 @@ final Set<String> skippedPaths = <String>{
   r'third_party/dart/tools', // not shipped in binary
   r'third_party/google_fonts_for_unit_tests', // only used in web unit tests
   r'third_party/gradle',
-  r'third_party/icu/filters',
-  r'third_party/icu/fuzzers',
-  r'third_party/icu/scripts',
-  r'third_party/icu/source/common/unicode/uvernum.h', // this file contains strings that confuse the analysis
-  r'third_party/icu/source/config',
-  r'third_party/icu/source/data/brkitr/dictionaries/burmesedict.txt', // explicitly handled by ICU license
-  r'third_party/icu/source/data/brkitr/dictionaries/cjdict.txt', // explicitly handled by ICU license
-  r'third_party/icu/source/data/brkitr/dictionaries/laodict.txt', // explicitly handled by ICU license
-  r'third_party/icu/source/data/dtd',
   r'third_party/java', // only used for Android builds
   r'third_party/libcxx/benchmarks',
   r'third_party/libcxx/docs',
@@ -187,7 +188,6 @@ final Set<String> skippedPaths = <String>{
   r'third_party/perfetto/src/trace_processor', // contains nothing that ends up in the binary executable
   r'third_party/perfetto/src/traced', // contains nothing that ends up in the binary executable
   r'third_party/perfetto/src/tracing', // contains nothing that ends up in the binary executable
-  r'third_party/protobuf', // build-time dependency only
   r'third_party/vulkan-deps/glslang/LICENSE', // excluded to make sure we don't accidentally apply it as a default license
   r'third_party/vulkan-deps/glslang/src/LICENSE.txt', // redundant with licenses inside files
   r'third_party/vulkan-deps/glslang/src/glslang/OSDependent/Web', // we only use glslang in impellerc, not in web apps
